+++ src/firmware-stm32/__app/inc/utility.h
@@ -17,7 +17,7 @@
 
 void assert2(const char *file, uint32_t line);
 
-void get_avg_from_circ(int last_idx, int ch_num, int avg_num, int bufflen, void* buff, int daq_bits, float* v1, float* v2, float* v3, float* v4, float* v5);
+void get_avg_from_circ(int last_idx, int ch_num, int avg_num, int bufflen, void* buff, int daq_bits, double* v1, double* v2, double* v3, double* v4, double* v5);
 int get_vcc_from_circ(int from, int total, int bufflen, int ch_num, int daq_bits, void* buff);
 int get_1ch_from_circ(int from, int total, int bufflen, int ch, int ch_num, int daq_bits, float vcc, float vref_cal, void* buff, uint8_t* out, int* idx);
 
@@ -29,7 +29,7 @@ static inline int get_last_circ_idx(int pos, int len)
     return last_idx;
 }
 
-float get_freq(int* prescaler, int* reload, int max_reload, int freq_osc, int freq_want);
+double get_freq(int* prescaler, int* reload, int max_reload, int freq_osc, int freq_want);
 
 void busy_wait(int us);
 float fastlog2 (float x);
@@ -41,6 +41,6 @@ void itoa_fast(char* s, int num, int radix);
 /* Author: Jakub Parez
  * Descr:  ultra fast float sprintf
  */
-void sprint_fast(char* s, const char* format, float fVal, int prec);
+void sprint_fast(char* s, const char* format, double fVal, int prec);
 
 #endif /* UTILS_H */
